# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "API Documentation",
        "contact": {
            "email": "ekoabdulgoffar129@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:49/bpm-backend-laravel/api/v1",
            "description": "API Documentation"
        }
    ],
    "paths": {
        "/api/v1/kriteria-survei": {
            "get": {
                "tags": [
                    "Kriteria Survei"
                ],
                "summary": "Get list of kriteria survei",
                "operationId": "3c2fdb3f0dadd317ba233a8b533029a4",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "ksr_id": {
                                                "type": "integer"
                                            },
                                            "ksr_nama": {
                                                "type": "string"
                                            },
                                            "ksr_status": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Kriteria Survei"
                ],
                "summary": "Create new kriteria survei",
                "operationId": "895a45769fd2f46bd898b2f493bd2125",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ksr_nama": {
                                        "type": "string"
                                    },
                                    "ksr_status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kriteria Survei created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/kriteria-survei/{id}": {
            "get": {
                "tags": [
                    "Kriteria Survei"
                ],
                "summary": "Get kriteria survei by ID",
                "operationId": "a02808ed3b19ae51cdfd62a531664951",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kriteria survei",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ksr_id": {
                                            "type": "integer"
                                        },
                                        "ksr_nama": {
                                            "type": "string"
                                        },
                                        "ksr_status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kriteria Survei not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Kriteria Survei"
                ],
                "summary": "Update kriteria survei",
                "operationId": "cc0eaa1f28878946bbbd1b8121d8d75b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kriteria survei",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ksr_nama": {
                                        "type": "string"
                                    },
                                    "ksr_status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kriteria Survei updated successfully"
                    },
                    "404": {
                        "description": "Kriteria Survei not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Kriteria Survei"
                ],
                "summary": "Delete kriteria survei",
                "operationId": "ed386e5eca838a1efb9dc472a959300c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kriteria survei",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kriteria Survei deleted successfully"
                    },
                    "404": {
                        "description": "Kriteria Survei not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token",
                "scheme": "bearer"
            }
        }
    },
...
}